requirements.txt
# ============================================================================
# requirements.txt
# ============================================================================
# Core dependencies
streamlit==1.31.0
pandas==2.2.0
openai==1.3.0
pydantic==2.6.0

# Async and HTTP
aiohttp==3.9.0
httpx==0.25.0
asyncio-throttle==1.0.2

# Data processing
openpyxl==3.1.2  # For Excel export
xlsxwriter==3.1.9  # For Excel writing
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0

# Optional but recommended
numpy==1.24.3
matplotlib==3.7.1
plotly==5.14.1